misc: attached gui cleanup
