+++ src/main/kotlin/me/tech/mcchestui/listeners/GUICloseListener.kt
@@ -1,7 +1,7 @@
 package me.tech.mcchestui.listeners
 
 import me.tech.mcchestui.GUI
-import me.tech.mcchestui.restoreAttachedInventory
+import me.tech.mcchestui.attached.restoreCachedInventory
 import org.bukkit.entity.Player
 import org.bukkit.event.EventHandler
 import org.bukkit.event.inventory.InventoryCloseEvent
@@ -19,7 +19,7 @@ internal class GUICloseListener(gui: GUI) : GUIEventListener(gui) {
             // TODO: remove listeners from child to reduce load.
             gui.attachedGui?.unregister()
 
-            restoreAttachedInventory(player as Player, gui)
+            restoreCachedInventory(player as Player, gui)
         }
 
         gui.onCloseInventory?.let { uiEvent ->
